fix: Handle OptimisticLockingFailureException on concurrent updates in UserService
Extracted the core `updateProfile` and `deleteAccount` logic into a new
`UserOperation` component with `@Idempotent` and `@Transactional(propagation = Propagation.REQUIRED)`.
Updated `UserService` to delegate to `UserOperation` and added `@Retryable`
annotations to automatically retry operations when an `ObjectOptimisticLockingFailureException` occurs.
This correctly separates transactional and retryable boundaries.
Unit tests have been updated and relocated to `UserOperationTest` to maintain AAA conventions.
