+++ src/main/java/com/tictactore/service/UserService.java
@@ -8,6 +8,9 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.retry.annotation.Backoff;
+import org.springframework.retry.annotation.Retryable;
+import org.springframework.orm.ObjectOptimisticLockingFailureException;
 
 import com.tictactore.dto.UpdateProfileRequest;
 import java.nio.charset.StandardCharsets;
@@ -34,6 +37,7 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final UserCreator userCreator;
+    private final UserOperation userOperation;
     private final ApplicationProperties properties;
     private final Clock clock;
     private final SecureRandom random = new SecureRandom();
@@ -164,83 +168,22 @@ public class UserService {
         }
     }
 
-    @Transactional
+    @Retryable(
+            retryFor = {ObjectOptimisticLockingFailureException.class},
+            maxAttempts = 3,
+            backoff = @Backoff(delay = 100)
+    )
     public User updateProfile(UUID userId, UpdateProfileRequest request) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new com.tictactore.exception.ResourceNotFoundException("User not found"));
-
-        if (request.getNickname() != null && !request.getNickname().trim().isEmpty()) {
-            String sanitized = sanitizeNickname(request.getNickname());
-            if (sanitized.isEmpty()) {
-                throw new com.tictactore.exception.ValidationException("Nickname cannot be empty");
-            }
-            if (!sanitized.equals(user.getNickname())) {
-                if (user.getLastNicknameUpdate() != null) {
-                    Instant nextAllowedUpdate = user.getLastNicknameUpdate().plus(30, ChronoUnit.DAYS);
-                    if (Instant.now(clock).isBefore(nextAllowedUpdate)) {
-                        throw new com.tictactore.exception.ValidationException("Nickname can only be changed once every 30 days");
-                    }
-                }
-                if (userRepository.existsByNickname(sanitized)) {
-                    throw new com.tictactore.exception.ValidationException("Nickname already taken");
-                }
-                user.setNickname(sanitized);
-                user.setLastNicknameUpdate(Instant.now(clock));
-            }
-        }
-
-        if (request.getLanguage() != null) {
-            String lang = request.getLanguage().toUpperCase();
-            if (!lang.equals("EN") && !lang.equals("DE")) {
-                throw new com.tictactore.exception.ValidationException("Language must be EN or DE");
-            }
-            user.setLanguage(lang);
-        }
-
-        if (request.getAvatar() != null) {
-            String avatarVal = request.getAvatar().trim();
-            if (!com.tictactore.validation.AvatarValidator.ALLOWED_AVATARS.contains(avatarVal)) {
-                throw new com.tictactore.exception.ValidationException("Invalid avatar selection");
-            }
-            user.setAvatar(avatarVal);
-        }
-
-        if (request.getTutorialCompleted() != null) {
-            user.setTutorialCompleted(request.getTutorialCompleted());
-        }
-
-        try {
-            return userRepository.save(user);
-        } catch (DataIntegrityViolationException e) {
-            throw new com.tictactore.exception.ValidationException("Nickname already taken");
-        }
+        return userOperation.updateProfile(userId, request);
     }
 
-    @Transactional
+    @Retryable(
+            retryFor = {ObjectOptimisticLockingFailureException.class},
+            maxAttempts = 3,
+            backoff = @Backoff(delay = 100)
+    )
     public void deleteAccount(UUID userId) {
-        if (userId == null) {
-            throw new IllegalArgumentException("User ID cannot be null");
-        }
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new com.tictactore.exception.ResourceNotFoundException("User not found"));
-
-        if (user.getProviderId() == null && "anonymous".equals(user.getAvatar())) {
-            return;
-        }
-
-        UUID uuid = java.util.UUID.randomUUID();
-        user.setEmail("deleted-" + uuid + "@tic-tac-tore.invalid");
-        user.setNickname("ex-player-" + uuid);
-        user.setAvatar("anonymous");
-        user.setProviderId(null);
-        user.setLanguage(null);
-        user.setLastNicknameUpdate(null);
-
-        try {
-            userRepository.flush();
-        } catch (org.springframework.orm.ObjectOptimisticLockingFailureException e) {
-            throw new IllegalStateException("Account was concurrently modified during deletion. Please try again.", e);
-        }
+        userOperation.deleteAccount(userId);
     }
     public UserPreferencesDto getLastRuleSystem() {
         return new UserPreferencesDto("STANDARD");
